move useUser
