+++ app/user/close/x/page.tsx
@@ -1,0 +1,5 @@
+import { Close } from '@/components/auth/OAuth';
+
+export default function ClosePage() {
+  return <Close />;
+}
